+++ app/javascript/components/ssr/feeds/ServerSideMemesPage.tsx
@@ -19,18 +19,16 @@ const ServerSideMemesPage: FunctionComponent<Props> = ({account, url, posts}: Pr
     const pathname = new URL(url).pathname;
     const history = createMemoryHistory({initialEntries: [pathname]});
     return (
-        <>
-            <div className="wrapper">
-                <Router history={history} location={pathname}>
-                    <NavBar showLogo={true} account={account}/>
-                    <div className="content">
-                        <h1 className={'page-heading'}>{I18n.t('ui.heading.top_posts')}</h1>
-                        <PostFeed account={account} sort={'new'} only={'memes'} posts={posts}/>
-                    </div>
-                    <Footer account={account}/>
-                </Router>
-            </div>
-        </>
+        <div className="wrapper">
+            <Router history={history} location={pathname}>
+                <NavBar showLogo={true} account={account}/>
+                <div className="content">
+                    <h1 className={'page-heading'}>{I18n.t('ui.heading.top_posts')}</h1>
+                    <PostFeed account={account} sort={'new'} only={'memes'} posts={posts}/>
+                </div>
+                <Footer account={account}/>
+            </Router>
+        </div>
     );
 };
