+++ app/javascript/components/ssr/feeds/ServerSideTopPage.tsx
@@ -19,18 +19,16 @@ const ServerSideHomePage: FunctionComponent<Props> = ({account, url, posts}: Pro
     const pathname = new URL(url).pathname;
     const history = createMemoryHistory({initialEntries: [pathname]});
     return (
-        <>
-            <div className="wrapper">
-                <Router history={history} location={pathname}>
-                    <NavBar showLogo={true} account={account}/>
-                    <div className="content">
-                        <h1 className={'page-heading'}>{I18n.t('ui.heading.top_posts')}</h1>
-                        <PostFeed account={account} sort={'best'} posts={posts}/>
-                    </div>
-                    <Footer account={account}/>
-                </Router>
-            </div>
-        </>
+        <div className="wrapper">
+            <Router history={history} location={pathname}>
+                <NavBar showLogo={true} account={account}/>
+                <div className="content">
+                    <h1 className={'page-heading'}>{I18n.t('ui.heading.top_posts')}</h1>
+                    <PostFeed account={account} sort={'best'} posts={posts}/>
+                </div>
+                <Footer account={account}/>
+            </Router>
+        </div>
     );
 };
